// src/app/models/historialclinico.interface.ts
export interface HistotialClinicoInterface {
    id?: string;
    idPatient?: string;
    Labios?: string;
    Carrillos?: string;
    Paladar?: string;
    Lengua?: string;
    PisodelaBoca?: string;
    Mucosa?: string;
    Encias?: string;
    Calculos?: string;
    Ganglios?: string;
    OtroMotivo?: string;
    complicacionesanestesia?: string;
    FiebreReumatica?: string;
    Diabetes?: string;
    Cardiopatias?: string;
    Tbc?: string;
    Artritis?: string;
    Hemorragias?: string;
    Neuralgias?: string;
    Sinusitiss?: string;
    Renales?: string;
    Alergias?: string;
    Otra?: string;
    Observaciones?: string;


}